generative-ai/requirement.txt
langchain
langchain_core
langchain-community
langgraph
langchain-openai
langchain-google-genai
langchain-groq
langchain-mistralai
python-dotenv
faiss-cpu
tiktoken
fastapi
uvicorn
requests

